tests/unit/test_detect.py
"""
Unit tests for detect module
"""
import pytest
import pandas as pd
from anomaly_log_detector.detect import BaselineParams


class TestBaselineParams:
    """Tests for BaselineParams dataclass"""

    def test_default_params(self):
        """Test default baseline parameters"""
        params = BaselineParams()
        assert params.window_size == 50
        assert params.stride == 25
        assert params.quantile == 0.95

    def test_custom_params(self):
        """Test custom baseline parameters"""
        params = BaselineParams(
            window_size=100,
            stride=50,
            quantile=0.99
        )
        assert params.window_size == 100
        assert params.stride == 50
        assert params.quantile == 0.99


class TestBaselineDetection:
    """Tests for baseline anomaly detection"""

    # Note: Full tests would require implementing the detection logic
    # or using test fixtures. This is a starting structure.

    def test_placeholder(self):
        """Placeholder test - implement actual detection tests"""
        # TODO: Add tests for baseline_detect function
        assert True
